Fully embrace profiles, creating an :inherited one
This should make the with-profile task work more sensibly. The
:inherited profile can be manipulated like any other. It can even appear
in [:profiles :inherited] instead of [:modules :inherited]. Either
should work. I toyed with prepending it to :default, but that can lead
to surprising results when :default is unmerged, depending on how your
profiles are organized. If desired, the user can always set [:profiles
:default] to include :inherited in project.clj.