Make an attempt at "compressing" expanded profiles
The latest leiningen expands profiles in the :active-profiles metadata,
thereby referring to profiles that may not exist in child projects to
which we'd like to activate the same profiles as in the parent. So we
try to compress them back into their original composite vectors. This is
not foolproof, however, as more complicated composite combinations could
result in the same expanded vector. To be continued...