Fix #19 by excluding :default from compression
Otherwise a child project may have different profiles in :default than
the parent, resulting in unpredictable active profiles when the modules
task is run.

It's interesting to note that it doesn't seem possible to fix this in
the opposite direction. A child cannot inherit a :default profile. It
must explicitly set it.