Update to work with latest lein, 2.5+
I think we're going to have to punt on being compatible with 2.5 due to
lein issue #1740, in which the order of more than a few merged profiles
can change.

Because lein's project ns now inits when it reads, we're subject to
merging profiles multiple times, so I changed the configless test to
account for that, alternatively with :replace metadata.