Account for different read behavior in lein 2.5+
The read function in 2.5 does a lot more than what it did in 2.4,
essentially now encapsulating the init-project function as well. But
this extra work greatly slows down projects with a lot of modules, so
this change uses read-raw instead of read when running in lein 2.5+.

2.5 can also be slower due to the new default value for :pedantic?,
which was nil in 2.4, so lein-modules users may wish to set it to false
in their modules' project.clj files, and it won't work in the :inherited
map.

This change also includes a minor new feature of running 'lein modules'
without any arguments just to see a listing of the modules in their
build order.