Discover child module versions automatically #29
This involved quite a bit of refactoring of the progeny stuff into
common so that versionization could use it, too. It looks for the
top-most project and then all of that project's children. This can lead
to an infinite loop when middleware is invoked so we introduce some
macros to disable middleware when parents and children are being read.

Unfortunately, discovering all the children can add quite a bit of time
to the build, almost double for some of the tasks in the Immutant build.

Parking on a branch for now.