Convert per-profile relative paths to source tree absolute paths
This fixes a "bug" (missing feature) in lein-modules where paths
inherited from the parent project are interpreted to be relative to
the inheriting child project, rather than being absolute with respect
to the project from which they are inherited.

This behavior made it difficult to have for instance a shared `/dev`
root source tree across all subproject in a repository.